// Microservice/frontend/src/types/auth/user.enum.ts
export enum UserTarget {
	LOSE_WEIGHT = 'LOSE WEIGHT',
	GAIN_WEIGHT = 'GAIN WEIGHT',
	IMPROVE_FLEXIBILITY = 'IMPROVE FLEXIBILITY',
	GENERAL_FITNESS = 'GENERAL FITNESS',
	BUILD_MUSCLE = 'BUILD MUSCLE',
	REHABILITATION_RECOVERY = 'REHABILITATION RECOVERY',
}

export enum UserPreferableActivity {
	YOGA = 'YOGA',
	CLIMBING = 'CLIMBING',
	STRENGTH_TRAINING = 'STRENGTH TRAINING',
	CROSSFIT = 'CROSSFIT',
	CARDIO_TRAINING = 'CARDIO TRAINING',
	REHABILITATION = 'REHABILITATION',
}

export type ActivityType = UserPreferableActivity;

export enum userWorkoutSearch {
	ALL = 'ALL',
	YOGA = 'YOGA',
	CLIMBING = 'CLIMBING',
	STRENGTH_TRAINING = 'STRENGTH TRAINING',
	CROSSFIT = 'CROSSFIT',
	CARDIO_TRAINING = 'CARDIO TRAINING',
	REHABILITATION = 'REHABILITATION',
}

export enum Role {
	CLIENT = 'CLIENT',
	ADMIN = 'ADMIN',
	COACH = 'COACH',
}